🚚(frontend) move type utils from utils to types directory
In order to improve project structure, type utils has been moved into `types`
directory. Previously type utils was in a file `types` within the `utils`
directory. By moving this file into `types`directory, all Richie's type
definitions are now in the same folder `types`.
