+++ src/frontend/js/data/useHistory/index.tsx
@@ -2,7 +2,7 @@ import { parse } from 'query-string';
 import React, { createContext, PropsWithChildren, useEffect, useState, useContext } from 'react';
 
 import { history, location } from 'utils/indirection/window';
-import { Maybe, Nullable } from 'utils/types';
+import { Maybe, Nullable } from 'types/utils';
 
 interface HistoryEntry {
   state: any;
